addr fb: best effort to swap back files after suite

+++ regression/multi-svc-app/multi_svc_app_suite_test.go
@@ -4,6 +4,7 @@
 package multi_svc_app_test
 
 import (
+       "errors"
        "fmt"
        "os"
        "testing"
@@ -43,14 +44,18 @@ var _ = AfterSuite(func() {
        _, err := toCLI.Run("app", "delete", "--yes")
        Expect(err).NotTo(HaveOccurred())
 
+       // Best-effort to swap back the files.
        for _, svcName := range []string{"front-end", "www", "back-end"} {
-              err := os.Rename(fmt.Sprintf("%s/main.go", svcName), fmt.Sprintf("%s/swap/main.go", svcName))
-              Expect(err).NotTo(HaveOccurred())
-              err = os.Rename(fmt.Sprintf("%s/swap/main.tmp", svcName), fmt.Sprintf("%s/main.go", svcName))
-              Expect(err).NotTo(HaveOccurred())
+              if _, err := os.Stat(fmt.Sprintf("%s/swap/main.tmp", svcName)); errors.Is(err, os.ErrNotExist) {
+                     // It's likely that a swap did not happen in the first place. No need to swap back.
+                     continue
+              }
+              _ = os.Rename(fmt.Sprintf("%s/main.go", svcName), fmt.Sprintf("%s/swap/main.go", svcName))
+              _ = os.Rename(fmt.Sprintf("%s/swap/main.tmp", svcName), fmt.Sprintf("%s/main.go", svcName))
        }
-       err = os.Rename("query/entrypoint.sh", "query/swap/entrypoint.sh")
-       Expect(err).NotTo(HaveOccurred())
-       err = os.Rename("query/swap/entrypoint.tmp", "query/entrypoint.sh")
-       Expect(err).NotTo(HaveOccurred())
+       if _, err := os.Stat("query/swap/entrypoint.tmp"); errors.Is(err, os.ErrNotExist) {
+              return
+       }
+       _ = os.Rename("query/entrypoint.sh", "query/swap/entrypoint.sh")
+       _ = os.Rename("query/swap/entrypoint.tmp", "query/entrypoint.sh")
 })
